Add files via upload
The Pine City Zoo App provides an animated map of the zoo, information of all the animals inhabiting the zoo & different activities that can be done at certain times. Users can also use the app to check the weather for the day they intend on visiting the zoo in order to plan their visit better and can provide feedback on their visit on the app.